Unified dialog parent routine
